authorization-service/requirements.txt
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.11.3'
autoflake==2.3.1; python_version >= '3.8'
black==24.8.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
cryptography==43.0.0; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.2.0; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
fastapi[standard]==0.112.0; python_version >= '3.8'
fastapi-cli[standard]==0.0.5
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
hiredis==3.0.0
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0
idna==3.7; python_version >= '3.5'
isort==5.13.2; python_full_version >= '3.8.0'
jinja2==3.1.4
joserfc==1.0.0; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==24.1; python_version >= '3.8'
passlib==1.7.4
pathspec==0.12.1; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
psycopg2==2.9.9; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.2
redis[hiredis]==5.0.8; python_version >= '3.7'
rich==13.7.1; python_full_version >= '3.7.0'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.32; python_version >= '3.7'
sqlmodel==0.0.21; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn[standard]==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
